Updated to Selenium 4.11
The code didn't work out of the box from master branch so I had to make
some updates. There were problems with the Duration class as well
but the main one was the driver managment. With Selenium 4.6.x that's
handled automatically but you will need to stop using the so blessed
3rd parity bonicarcia WebDriverManager (R.I.P. you have been on great
help in the past)